Need to flesh it out a bit more but timers available. Need to be able to add new ones, clean up how it looks

+++ src/util/db-setupd.ts
@@ -66,8 +66,8 @@ export const setup = (db: sqlite.Database): void => {
       CREATE TABLE IF NOT EXISTS timer (
         id INTEGER NOT NULL,
         name TEXT NOT NULL,
-        minTime	INTEGER,
-        maxTime	INTEGER,
+        minTime	NUMERIC,
+        maxTime	NUMERIC,
         measurement TEXT,
         recipeId INTEGER,
         FOREIGN KEY(recipeId)
